// apps/api/src/instagram/entities/instagram-insights.ts
export enum InstagramInsightsMetrics {
    impressions = 'impressions',
    reach = 'reach',
    follower_count = 'follower_count',
    email_contacts = 'email_contacts',
    phone_call_clicks = 'phone_call_clicks',
    text_message_clicks = 'text_message_clicks',
    get_directions_clicks = 'get_directions_clicks',
    website_clicks = 'website_clicks',
    profile_views = 'profile_views',
    audience_gender_age = 'audience_gender_age',
    audience_locale = 'audience_locale',
    audience_country = 'audience_country',
    audience_city = 'audience_city',
    online_followers = 'online_followers',
    accounts_engaged = 'accounts_engaged',
    total_interactions = 'total_interactions',
    likes = 'likes',
    comments = 'comments',
    shares = 'shares',
    saves = 'saves',
    replies = 'replies',
    engaged_audience_demographics = 'engaged_audience_demographics',
    reached_audience_demographics = 'reached_audience_demographics',
    follower_demographics = 'follower_demographics',
    follows_and_unfollows = 'follows_and_unfollows',
    profile_links_taps = 'profile_links_taps',
  }